Decouples bastion from GCP.
